refactor: reduce hierarchy and merge related files
License: MIT
