+++ src/argv-files-handler.js
@@ -1,6 +1,6 @@
 import { app } from 'electron'
 import fs from 'fs-extra'
-import { addToIpfs } from '../utils'
+import addToIpfs from './add-to-ipfs'
 
 export default async function (ctx) {
   const handleArgv = async argv => {
@@ -20,15 +20,6 @@ export default async function (ctx) {
   // Works for Windows context menu
   app.on('second-instance', (_, argv) => {
     handleArgv(argv)
-  })
-
-  // macOS tray drop files
-  ctx.tray.on('drop-files', async (_, files) => {
-    for (const file of files) {
-      await addToIpfs(ctx, file)
-    }
-
-    ctx.launchWebUI('/files', { focus: false })
   })
 
   // Checks current proccess
