+++ src/auto-updater/index.js
@@ -1,6 +1,8 @@
 import { autoUpdater } from 'electron-updater'
 import i18n from 'i18next'
-import { logger, notify, quitAndInstall } from '../utils'
+import quitAndInstall from './quit-and-install'
+import logger from '../common/logger'
+import notify from '../common/notify'
 
 let userRequested = false
