--- src/early/index.js
@@ -1,6 +1,0 @@
-import protocolHandlers from './protocol-handlers'
-
-// Functions that must run on app 'will-finish-launching'.
-export default function (ctx) {
-  protocolHandlers(ctx)
-}
