+++ src/exec-or-sudo.js
@@ -1,9 +1,9 @@
 import i18n from 'i18next'
+import util from 'util'
 import sudo from 'sudo-prompt'
 import { dialog, app } from 'electron'
-import logger from './logger'
-import { recoverableErrorDialog } from '../dialogs'
-import util from 'util'
+import { recoverableErrorDialog } from './dialogs'
+import logger from './common/logger'
 
 const execFile = util.promisify(require('child_process').execFile)
