+++ src/ipfs-on-path/scripts/install.ps1
