+++ src/ipfs-on-path/scripts/uninstall.js
@@ -1,5 +1,5 @@
 const { DEST_SCRIPT } = require('./consts')
-const { revert } = require('../../../utils/scripts/backup')
+const { revert } = require('./backup')
 const argv = require('yargs').argv
 
 revert(argv.data, 'ipfs', DEST_SCRIPT)
