+++ src/ipfs-on-path/scripts/uninstall.ps1
