+++ src/take-screenshot.js
@@ -1,6 +1,9 @@
 import i18n from 'i18next'
 import { clipboard, nativeImage, ipcMain } from 'electron'
-import { notify, notifyError, logger, setupGlobalShortcut, IS_MAC } from '../utils'
+import logger from './common/logger'
+import { IS_MAC } from './common/consts'
+import { notify, notifyError } from './common/notify'
+import setupGlobalShortcut from './setup-global-shortcut'
 
 const CONFIG_KEY = 'screenshotShortcut'
