+++ src/webui/preload.js
@@ -1,11 +1,11 @@
-const { ipcRenderer, remote } = require('electron')
-const screenshotHook = require('./screenshot')
-const connectionHook = require('./connection-status')
 const toPull = require('stream-to-pull-stream')
+const { ipcRenderer, remote } = require('electron')
 const readdir = require('recursive-readdir')
 const fs = require('fs-extra')
 const path = require('path')
-const pkg = require('../../../package.json')
+const screenshotHook = require('./screenshot')
+const connectionHook = require('./connection-status')
+const pkg = require('../../package.json')
 
 const COUNTLY_KEY = '47fbb3db3426d2ae32b3b65fe40c564063d8b55d'
 const COUNTLY_KEY_TEST = '6b00e04fa5370b1ce361d2f24a09c74254eee382'
