+++ src/webui/screenshot.js
